git-buildpackage Debian release 0.6.9

Conflicts:
	debian/changelog

+++ tests/component/__init__.py
@@ -24,8 +24,8 @@ import os
 import shutil
 import tempfile
 from StringIO import StringIO
-
 from nose import SkipTest
+from nose.tools import eq_, ok_     # pylint: disable=E0611
 
 import gbp.log
 from  gbp.git import GitRepository, GitRepositoryError
@@ -89,7 +89,8 @@ class ComponentTestBase(object):
     def teardown_class(cls):
         """Test class case teardown"""
         # Return original environment
-        os.environ = cls.orig_env
+        os.environ.clear()
+        os.environ.update(cls.orig_env)
 
     def __init__(self):
         """Object initialization"""
@@ -111,7 +112,8 @@ class ComponentTestBase(object):
         """Test case teardown"""
         # Restore original working dir
         os.chdir(self._orig_dir)
-        shutil.rmtree(self._tmpdir)
+        if not os.getenv("GBP_TESTS_NOCLEAN"):
+            shutil.rmtree(self._tmpdir)
 
         self._capture_log(False)
 
@@ -119,9 +121,12 @@ class ComponentTestBase(object):
     def _check_repo_state(cls, repo, current_branch, branches, files=None):
         """Check that repository is clean and given branches exist"""
         branch = repo.branch
-        assert branch == current_branch
-        assert repo.is_clean()
-        assert set(repo.get_local_branches()) == set(branches)
+        eq_(branch, current_branch)
+        ok_(repo.is_clean())
+        local_branches = repo.get_local_branches()
+        assert_msg = "Branches: expected %s, found %s" % (branches,
+                                                          local_branches)
+        eq_(set(local_branches), set(branches), assert_msg)
         if files is not None:
             # Get files of the working copy recursively
             local = set()
@@ -136,9 +141,9 @@ class ComponentTestBase(object):
                     local.add(os.path.relpath(os.path.join(dirpath, dirname),
                                               repo.path) + '/')
             extra = local - set(files)
-            assert not extra, "Unexpected files in repo: %s" % list(extra)
+            ok_(not extra, "Unexpected files in repo: %s" % list(extra))
             missing = set(files) - local
-            assert not missing, "Files missing from repo: %s" % list(missing)
+            ok_(not missing, "Files missing from repo: %s" % list(missing))
 
     def _capture_log(self, capture=True):
         """ Capture log"""
@@ -163,10 +168,10 @@ class ComponentTestBase(object):
     def _check_log(self, linenum, string):
         """Check that the specified line on log matches expectations"""
         if self._log is None:
-            assert False, "BUG in unittests: no log captured!"
+            raise Exception("BUG in unittests: no log captured!")
         output = self._get_log()[linenum].strip()
-        assert output.startswith(string), ("Expected: '%s...' Got: '%s'" %
-                                           (string, output))
+        ok_(output.startswith(string), ("Expected: '%s...' Got: '%s'" %
+                                        (string, output)))
 
     def _clear_log(self):
         """Clear the mock strerr"""
